requirements.txt
fastmcp>=2.0.0
python-dotenv>=1.0.0
oracledb>=2.0.0  # optional — Oracle 사용 시만 필요
